Classes MVP without Frontend/CLI/HNSW
