--- data/drift_log.csv
@@ -1,29 +1,0 @@
-ID,TIMESTAMP,I/O TYPE,SEMANTIC ROLLING COS,SEMANTIC TRAINING COS,CONCEPT ROLLING COS,CONCEPT TRAINING COS
-f93cbab3-e976-4c04-827a-3e461c768d05,2025-03-27T06:41:35.780Z,input,1.0000000000000002,1.0000000000000002,1,1
-f93cbab3-e976-4c04-827a-3e461c768d05,2025-03-27T06:41:35.780Z,output,1,1,1,1
-175f3710-6d5f-4b23-a217-9b68e5cdd978,2025-03-27T07:08:45.834Z,input,0.8954042456354891,0.8954042456354891,0.9887458291120869,0.9887458291120869
-175f3710-6d5f-4b23-a217-9b68e5cdd978,2025-03-27T07:08:45.834Z,output,0.895713213217582,0.895713213217582,0.9893786036567933,0.9893786036567933
-f6cae3e0-1524-48b3-9d90-aff1c78c11db,2025-03-27T07:08:52.008Z,input,0.9490311594952278,0.9490311594952278,0.9937315448285899,0.9937315448285899
-f6cae3e0-1524-48b3-9d90-aff1c78c11db,2025-03-27T07:08:52.008Z,output,0.9491421469538633,0.9491421469538633,0.9940806043926216,0.9940806043926216
-fe8f074c-ae35-458b-aade-4ea7b404168a,2025-03-27T07:08:53.008Z,input,0.9703004456807715,0.9703004456807715,0.9960158586004877,0.9960158586004877
-fe8f074c-ae35-458b-aade-4ea7b404168a,2025-03-27T07:08:53.008Z,output,0.9703512455826723,0.9703512455826723,0.9962362081367161,0.9962362081367161
-adde06f4-663a-4dfc-b9fd-65d79a67b9d0,2025-03-27T07:08:53.882Z,input,0.9806646856417617,0.9806646856417617,0.9972471119096632,0.9972471119096632
-adde06f4-663a-4dfc-b9fd-65d79a67b9d0,2025-03-27T07:08:53.882Z,output,0.980691908050669,0.980691908050669,0.9973986108048084,0.9973986108048084
-7715aeb5-3afa-4121-8fef-3a2d40bbc9c5,2025-03-27T07:08:54.758Z,input,0.9864436965499939,0.9864436965499939,0.997985123631888,0.997985123631888
-7715aeb5-3afa-4121-8fef-3a2d40bbc9c5,2025-03-27T07:08:54.758Z,output,0.9864599401619532,0.9864599401619532,0.998095592698213,0.998095592698213
-662ddb32-6255-4a90-a566-d81a5b5f7f08,2025-03-27T07:08:55.624Z,input,0.9899813561150311,0.9899813561150311,0.9984619067957321,0.9984619067957321
-662ddb32-6255-4a90-a566-d81a5b5f7f08,2025-03-27T07:08:55.624Z,output,0.9899918251567404,0.9899918251567404,0.99854598932353,0.99854598932353
-652067b9-8cb6-42fa-864e-2035c1190f8f,2025-03-27T07:08:56.466Z,input,0.9922994917211683,0.9922994917211683,0.9987875801199385,0.9987875801199385
-652067b9-8cb6-42fa-864e-2035c1190f8f,2025-03-27T07:08:56.466Z,output,0.9923066408151859,0.9923066408151859,0.99885370439067,0.99885370439067
-35cae0ac-0540-404a-b452-097859837b8b,2025-03-27T07:08:57.301Z,input,0.9938990722823914,0.9938990722823914,0.9990198305115159,0.9990198305115159
-35cae0ac-0540-404a-b452-097859837b8b,2025-03-27T07:08:57.301Z,output,0.993904177624692,0.993904177624692,0.9990731860160892,0.9990731860160892
-6dd409ca-9f60-48ff-ba67-9267da2610cd,2025-03-27T07:09:01.231Z,input,0.9938990722823914,0.9950484702188936,0.9990198305115159,0.9991912398375996
-6dd409ca-9f60-48ff-ba67-9267da2610cd,2025-03-27T07:09:01.231Z,output,0.993904177624692,0.9950522489054736,0.9990731860160892,0.9992351947293165
-44735651-e0dc-40fb-a558-0ca8d5001d2b,2025-03-27T07:09:01.994Z,input,1,0.9959017483138545,1.0000000000000002,0.9993213359849556
-44735651-e0dc-40fb-a558-0ca8d5001d2b,2025-03-27T07:09:01.994Z,output,1,0.9959046272036327,0.9999999999999998,0.9993581708239917
-573f2d35-2256-45b5-b5cf-626394562969,2025-03-27T07:09:02.729Z,input,1,0.9965523851641482,1.0000000000000002,0.9994223992365392
-573f2d35-2256-45b5-b5cf-626394562969,2025-03-27T07:09:02.729Z,output,1,0.9965546323540249,0.9999999999999998,0.9994537128045704
-6f90a477-b624-4778-bcc7-8b9ae09d4818,2025-03-27T07:09:03.517Z,input,1,0.997059732895389,1.0000000000000002,0.9995024652865339
-6f90a477-b624-4778-bcc7-8b9ae09d4818,2025-03-27T07:09:03.517Z,output,1,0.9970615228231595,0.9999999999999998,0.9995294114013288
-83111c01-4554-4135-96f7-2171c1a05601,2025-03-27T07:09:04.251Z,input,1,0.9974629282811313,1.0000000000000002,0.9995669710096596
-83111c01-4554-4135-96f7-2171c1a05601,2025-03-27T07:09:04.251Z,output,1,0.9974643783561037,0.9999999999999998,0.9995904031693741
