--- util/7-printLogCLI.js
@@ -1,7 +1,0 @@
-// TODO: Need to make the logic to:
-// 1. Get the data from the CLI log file
-// 2. Filter it to show the relevant time period
-// 3. compute some % to show # of drifts during that time period
-// 4. print to console
-// ? Should we make a log of CLI queries? 
-// Do we want to keep track of previous results
